Extract data access layer
