+++ webapp/src/app/page.tsx
@@ -1,9 +1,6 @@
-import { Cache } from '@/Cache';
+import { accessProjects } from '@/dataAccess';
 import HomeDashboard from '@/components/HomeDashboard';
-import MessageAdapterFactory from '@/utils/adapters/MessageAdapterFactory';
 import { ProjectCardProps } from '@/components/ProjectCard';
-import { RepoGit } from '@/RepoGit';
-import { ServerConfig } from '@/utils/serverConfig';
 
 // Force dynamic rendering for this page. By default Next.js attempts to render
 // this page statically. That means that it tries to render the page at build
@@ -19,27 +16,8 @@ import { ServerConfig } from '@/utils/serverConfig';
 export const dynamic = 'force-dynamic';
 
 export default async function Home() {
-  const serverConfig = await ServerConfig.read();
-  const projects = await new Promises(
-    serverConfig.projects.map(async (project) => {
-      await RepoGit.cloneIfNotExist(project);
-      const repoGit = await RepoGit.getRepoGit(project);
-      const lyraConfig = await repoGit.getLyraConfig();
-      const projectConfig = lyraConfig.getProjectConfigByPath(
-        project.projectPath,
-      );
-      const msgAdapter = MessageAdapterFactory.createAdapter(projectConfig);
-      const messages = await msgAdapter.getMessages();
-      const store = await Cache.getProjectStore(projectConfig);
-      const languagesWithTranslations = projectConfig.languages.map(
-        async (lang) => {
-          const translations = await store.getTranslations(lang);
-          return { lang, translations };
-        },
-      );
-      return { languagesWithTranslations, messages, name: project.name };
-    }),
-  )
+  const projectData = await accessProjects();
+  const projects = await new Promises(projectData)
     .map<ProjectCardProps>(async (value) => {
       const { name, messages, languagesWithTranslations } = value;
       const languages = await new Promises(languagesWithTranslations)
